# docs/openapi.yaml
openapi: 3.1.0
info:
  title: addons-core — Stremio-Addon Aggregator & OMSS API
  version: 1.0.0
  description: Production-grade Stremio-addon aggregator exposing an OMSS-compliant streaming backend, progressive waterfall scraping, dynamic addon management, debrid caching, and health observability.
servers:
  - url: http://localhost:3006
    description: addons-core API Server
...
paths:
...
  /v1/catalogs:
    get:
      tags:
        - Catalogs
      summary: List catalog descriptors from enabled catalog-capable addons (Phase 12 §15.1)
      responses:
        '200':
          description: Catalog descriptor list
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/catalogs/{slug}/{type}/{catalogId}:
    get:
      tags:
        - Catalogs
      summary: Fetch one normalized, cached catalog page from an addon (Phase 12 §15.1)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: catalogId
          in: path
          required: true
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: genre
          in: query
          schema:
            type: string
        - name: skip
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Normalized catalog page
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          description: Catalog upstream failed
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or API Key
    SessionAuth:
      type: apiKey
      in: cookie
      name: cineflix_session
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid parameter
            details:
              type: array
              items:
                type: object
        requestId:
          type: string
          example: req-abc123
        revision:
          type: integer
          example: 5
  responses:
...
    Unauthorized:
      description: Authentication token missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource or media not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...